requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
redis[asyncio]==5.2.0
sqlalchemy[asyncio]==2.0.36
anthropic==0.40.0
openai==1.57.0
pgvector==0.4.2
sentence-transformers==3.3.1
structlog==24.4.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.0
slowapi==0.1.9
alembic==1.18.4
pytest==8.3.4
pytest-asyncio==0.24.0
